+++ src/pygnssutils/gnssmqttclient.py
@@ -6,10 +6,18 @@ source and (optionally) sending the data to a designated writeable output
 medium (serial, file, socket, queue).
 
 Calling app, if defined, can implement the following methods:
-- set_event() - create <<spartn_read>> event
-- dialog() - return reference to MQTT client configuration dialog
+ - set_event() - create <<spartn_read>> event
+ - dialog() - return reference to MQTT client configuration dialog
 
-Thingstream > Location Services > PointPerfect Thing > Credentials
+Can utilise the following environment variables:
+ - MQTTKEY - SPARTN payload decription key (valid for 4 weeks)
+ - MQTTCRT - MQTT server (PointPerfect) TLS certificate
+ - MQTTPEM - MQTT server (PointPerfect) TLS key
+ - MQTTCLIENTID - MQTT server client ID
+
+Credentials can be download from:
+ Thingstream > Location Services > PointPerfect Thing > Credentials
+
 Default location for key files is user's HOME directory
 
 Created on 20 Feb 2023
@@ -86,8 +94,14 @@ class GNSSMQTTClient:
             "topic_ip": 1,
             "topic_mga": 1,
             "topic_key": 1,
-            "tlscrt": path.join(Path.home(), f"device-{clientid}-pp-cert.crt"),
-            "tlskey": path.join(Path.home(), f"device-{clientid}-pp-key.pem"),
+            "tlscrt": getenv(
+                "MQTTCRT",
+                default=path.join(Path.home(), f"device-{clientid}-pp-cert.crt"),
+            ),
+            "tlskey": getenv(
+                "MQTTPEM",
+                default=path.join(Path.home(), f"device-{clientid}-pp-key.pem"),
+            ),
             "spartndecode": 0,
             "spartnkey": getenv("MQTTKEY", default=None),
             "spartnbasedate": datetime.now(timezone.utc),