+++ tests/dummysocket.py
@@ -10,20 +10,31 @@ Created on 21 Aug 2024
 :license: BSD 3-Clause
 """
 
-DEFAULT_BUFSIZE = 4096
-
 
 class DummySocket:
     """
     Dummy socket class for testing SocketWrapper.
     """
 
-    def __init__(self, filename: str, bufsize: int = DEFAULT_BUFSIZE):
+    def __init__(
+        self,
+        filename: str,
+        bufsize: int = 4096,
+        timeout: bool = False,
+    ):
         """
         Constructor.
+
+        Reads binary data from a file into a working buffer,
+        representing the socket stream.
+
+        :param str filename: filename
+        :param int bufsize: size of buffer
+        :param bool timeout: simulate TimeoutError?
         """
 
         self._buffer = b""
+        self._timeout = timeout
         with open(filename, "rb") as infile:
             while len(self._buffer) < bufsize:
                 b = infile.read(16)
@@ -33,16 +44,26 @@ class DummySocket:
 
     def recv(self, n: int) -> bytes:
         """
-        Read n bytes from dummy socket.
+        Receive n bytes from dummy socket.
+
+        :param int n: number of bytes to read
+        :return: bytes read
+        :rtype: bytes
         """
 
+        if self._timeout:
+            raise TimeoutError("simulated TimeoutError")
         b = self._buffer[:n]
         self._buffer = self._buffer[n:]
         return b
 
-    def send(self, data) -> int:
+    def send(self, data: bytes) -> int:
         """
-        Send data to socket
+        Send data to socket.
+
+        :param bytes data: data to send
+        :return: number of bytes sent
+        :rtype: int
         """
 
         print(f"data sent: {data}")
@@ -50,13 +71,12 @@ class DummySocket:
 
     def sendall(self, data):
         """
-        Send data to socket
+        Sendall data to socket.
         """
 
         print(f"data sent: {data}")
-        return None
 
-    def close():
+    def close(self):
         """
         Close socket
         """